services/ble_collector/requirements.txt
python-dotenv
bleak
influxdb_client
